+++ LibPkmGC/include/LibPkmGC/Colosseum/SaveEditing/Save.h
@@ -42,6 +42,8 @@ public:
 
 	void save(void);
 
+	void fixBugsAffectingPokemon(void);
+
 protected:
 	void loadFields(void);