+++ LibPkmGC/src/LibPkmGC/Colosseum/Common/PlayerData.cpp
@@ -48,8 +48,8 @@ void PlayerData::loadFields(void) {
 	LD_FIELD_E(u8, trainerGender, 0xa80, Gender);
 	if (trainerGender > Female) trainerGender = Male;
 
-	LD_FIELD(u32, money, 0xa84);
-	LD_FIELD(u32, pkCoupons, 0xa88);
+	LD_FIELD_MAX(u32, money, 0xa84, 9999999);
+	LD_FIELD_MAX(u32, pkCoupons, 0xa88, 9999999);
 
 	LD_SUBSTRUCTURE_ARRAY(Pokemon, party, 6, 0x30);
 	LD_SUBSTRUCTURE(BagData, bag, 0x780);
@@ -59,8 +59,8 @@ void PlayerData::save(void) {
 	GC::PlayerData::save();
 	if (trainerGender > Female) trainerGender = Male;
 	SV_FIELD_E(u8, trainerGender, 0xa80, Gender);
-	SV_FIELD(u32, money, 0xa84);
-	SV_FIELD(u32, pkCoupons, 0xa88);
+	SV_FIELD_MAX(u32, money, 0xa84, 9999999);
+	SV_FIELD_MAX(u32, pkCoupons, 0xa88, 9999999);
 	SV_FIELD(u32, pkCoupons, 0xa8c);
 
 	SV_SUBSTRUCTURE_ARRAY(Pokemon, party, 6, 0x30);